+++ src/Framework/ContentPacks/Model/Content.cs
@@ -14,6 +14,7 @@ namespace QuestFramework.Framework.ContentPacks.Model
         public ISemanticVersion Format { get; set; }
         public List<QuestData> Quests { get; set; }
         public List<QuestOffer<JObject>> Offers { get; set; } = new List<QuestOffer<JObject>>();
+        public List<CustomBoardData> CustomBoards { get; set; } = new List<CustomBoardData>();
 
         internal IContentPack Owner { get; set; }
 
@@ -24,6 +25,7 @@ namespace QuestFramework.Framework.ContentPacks.Model
                 Format = this.Format,
                 Quests = this.Quests.Select(q => q.Translate(translation)).ToList(),
                 Offers = this.Offers.Select(o => TranslationUtils.TranslateOffer(translation, o)).ToList(),
+                CustomBoards = this.CustomBoards.Select(b => TranslationUtils.Translate(translation, b)).ToList(),
                 Owner = this.Owner,
             };
         }