+++ src/Framework/ContentPacks/Model/QuestData.cs
@@ -42,10 +42,7 @@ namespace QuestFramework.Framework.ContentPacks.Model
 
         public QuestData Translate(ITranslationHelper translation)
         {
-            var toTranslate = JObject.FromObject(this);
-
-            return TranslationUtils.Translate(translation, toTranslate)
-                .ToObject<QuestData>();
+            return TranslationUtils.Translate(translation, this);
         }
     }
 }