+++ docs/changelog.md
@@ -4,12 +4,18 @@
 
 ### Common changes
 
+- Added support for custom quest boards (in any location what you want)
 - Managed quests are scrollable in questlog if content needs scroll (like special orders)
 - Added support for multiple objectives
 - Fixed some multiplayer problems
 
+### For Content Pack modders
+
+- Added field `CustomBoards` for define custom quest boards (see docs)
+
 ### For SMAPI modders
 
+- New managed API method `RegisterCustomBoard`
 - New CustomQuest apis: `GetCurrentObjectives`, `UpdateUpdateCurrentObjectives` methods and property `Objectives` for define multiple objectives for a quest. (You must handle objectives manually in your quest code)
 
 ## 1.2.1